Code get_daily and get_annual function in the simulation children. Create Weather class with a one year prediction model randomly generated

+++ src/Central.java
@@ -2,7 +2,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Central extends Producer {
-    Central(double prod) {
-        super(new Type1(new ArrayList<Double>(List.of(prod))));
+    Central(float prod) {
+        super(new Type1(new ArrayList<Float>(List.of(prod))));
     }
 }
